requirements.txt
absl-py==2.1.0
aiosqlite==0.20.0
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
astunparse==1.6.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
distlib==0.3.8
Django==5.0.3
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.111.0
fastapi-cli==0.0.3
filelock==3.13.4
flatbuffers==24.3.25
fonttools==4.51.0
fsspec==2024.3.1
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.62.2
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
kares==0.12
keras==3.3.3
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
MyFirstProject==0.0.1
namex==0.0.8
networkx==3.3
numpy==1.26.4
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
orjson==3.10.3
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pipenv==2024.1.0
platformdirs==4.2.1
protobuf==4.25.3
psutil==5.9.8
psycopg2-binary==2.9.9
py-cpuinfo==9.0.0
pydantic==2.7.1
pydantic-settings==2.3.4
pydantic_core==2.18.2
pygame==2.5.2
Pygments==2.18.0
PyMySQL==1.1.1
pyparsing==3.1.2
pyTelegramBotAPI==4.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
python-telegram-bot==21.6
pytz==2024.1
PyYAML==6.0.1
requests==2.32.0
rich==13.7.1
scipy==1.13.0
seaborn==0.13.2
setuptools==70.0.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.30
sqlparse==0.5.0
starlette==0.40.0
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
termcolor==2.4.0
thop==0.1.1.post2209072238
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.3
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.10.0
ultralytics==8.2.2
urllib3==2.2.2
uvicorn==0.29.0
uvloop==0.19.0
virtualenv==20.26.1
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.2
wheel==0.43.0
wrapt==1.16.0

djangorestframework~=3.15.2
cryptography~=43.0.3